+++ lib/bloom_remit_client/requests/recipients/create.rb
@@ -3,6 +3,8 @@ module BloomRemitClient
     module Recipients
       class Create < Base
 
+        include Concerns::HasBaseAuthentification
+
         PATH = "/api/v1/partners/:api_token/senders/:sender_id/recipients.json"
 
         attribute :recipient, Hash