+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Add a way to get the deposit strategies. `slug` and `name` for now
+
 ## [0.7.1]
 ### Fixed
 - Factory `bloom_remit_client_create_payment_response` renamed to `bloom_remit_client_responses_payments_create`